+++ src/main/java/chess/controller/ChessController.java
@@ -3,8 +3,8 @@ package chess.controller;
 import chess.domain.ChessGame;
 import chess.domain.board.Board;
 import chess.domain.command.Command;
-import chess.domain.dto.BoardDto;
-import chess.domain.dto.BoardStatusDto;
+import chess.dto.BoardDto;
+import chess.dto.BoardStatusDto;
 import chess.exception.GameIsNotStartException;
 import chess.view.InputView;
 import chess.view.OutputView;
