+++ src/main/java/chess/dto/MovablePositionDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 public class MovablePositionDto {
     private String target;
