+++ src/main/java/chess/dto/MoveRequestDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 public class MoveRequestDto {
     private String target;
