+++ src/main/resources/templates/main.html
@@ -5,7 +5,8 @@
     <title>체스 홈</title>
 </head>
 <body>
-<form action="/game">
+<form action="/start">
+    방번호 : <input type="text" name="room">
     <input type="submit" value="입장하기">
 </form>
 </body>
